Marketing-Content-Pack-main/requirements.txt
fastapi
uvicorn
stripe
fpdf
python-dotenv
requests
openai>=1.0.0,<2.0.0
python-dotenv
sendgrid
